migration: move migration code from libvirt to migration.py
As part of refactor, move migration code from libvirt to migration.py
to make the framework cleaner and adopt the changes in env_process to
use from migration.py
